added everything for the first release. game now starts without any errors.

+++ src/main/java/com/itskillerluc/duclib/data/animation/serializers/Animation.java
@@ -4,8 +4,8 @@ import com.google.gson.JsonParseException;
 
 import java.util.Map;
 
-public record Animation(boolean loop, int animationLength, float speed, Map<String, Bone> bones) {
-    public Animation(boolean loop, int animationLength, float speed, Map<String, Bone> bones) {
+public record Animation(boolean loop, float animationLength, float speed, Map<String, Bone> bones) {
+    public Animation(boolean loop, float animationLength, float speed, Map<String, Bone> bones) {
         if (bones == null) {
             throw new JsonParseException("field bones is missing");
         }
